backend/kgc/pyproject.toml
[project]
name = "foodatlas-kgc"
version = "0.1.0"
description = "FoodAtlas knowledge graph construction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "click>=8.1.0",
    "biopython>=1.84",
    "inflection>=0.5.1",
    "tqdm>=4.66.0",
    "pympler>=1.0",
    "pydantic-settings>=2.2.0",
    "owlready2>=0.46",
    "rdflib>=7.0.0",
    "lxml>=5.1.0",
    "openpyxl>=3.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.25.0",
    "xmltodict>=0.13.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.2.2",
]

[tool.uv]
python-preference = "system"
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "bandit>=1.8.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
]

# PYTEST
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
filterwarnings = ["error", "ignore::DeprecationWarning"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
